+++ selene-lib/tests/lints/standard_library/constants.lua
@@ -5,3 +5,5 @@ collectgarbage("doge")
 collectgarbage(1)
 
 collectgarbage("coun" .. "t")
+
+collectgarbage()